+++ CHANGELOG.md
@@ -2,8 +2,6 @@
 -----
 
 *   Updates:
-    *   Add theme support to Up Next screen
-        ([#1605](https://github.com/Automattic/pocket-casts-android/pull/1605))
     *   Save selected tab on Player screen after screen rotation
         ([#1598](https://github.com/Automattic/pocket-casts-android/issues/1598))
 
@@ -20,9 +18,14 @@
         ([#1596](https://github.com/Automattic/pocket-casts-android/pull/1596))
     *   Fix text being cut off in discover search bar at high zoom 
         ([#1601](https://github.com/Automattic/pocket-casts-android/pull/1601))
+    *   Fix issues where multi-select on one screen would affect another screen
+        ([#1579](https://github.com/Automattic/pocket-casts-android/pull/1579),
+         [#1580](https://github.com/Automattic/pocket-casts-android/pull/1580))
 *   Updates:
     *   Display dynamic colors for widget
         ([#1588](https://github.com/Automattic/pocket-casts-android/pull/1588))
+    *   Add theme support to Up Next screen
+        ([#1605](https://github.com/Automattic/pocket-casts-android/pull/1605))
 
 7.53
 -----